// dynamite/src/interfaces/IDynamiteCompany.ts
export interface IDynamiteCompany {
    name: string;
    companyUrl: string;
    usernameLow: string;
    icon: {
        hasIcon: boolean;
        hasWebp: boolean;
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xsw: string;
        smw: string;
        mdw: string;
        lgw: string;
    }
}